pyproject.toml
[project]
name = "programering-inom-data-plattform-testar-koden"
version = "0.1.0"
description = "Riksdagsdata projekt"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "uvicorn",
    "httpx",
    "pandas>=1.4.0,<3",
    "aiokafka>=0.13.0",
    "kafka-python>=2.3.0",
    "jupyter>=1.1.1",
    "notebook>=7.5.5",
    "jupyterlab>=4.5.6",
    "matplotlib>=3.10.8",
    "seaborn>=0.13.2",
    "psycopg2-binary>=2.9.11",
    "streamlit==1.41.0",
    "plotly>=6.6.0",
    "altair==5.5.0",
    "sqlalchemy>=2.0.48",
]

[tool.uv]
package = false
